+++ sandbox/common/run.js
@@ -7,11 +7,12 @@ const defaults = {
 async function run(init, options = {}) {
   const props = { ...defaults, ...options };
   const canvas = document.getElementById('canvas');
-  const vis = await new window.Visualization({
+
+  window.visualisation = await new window.Visualization({
     canvas,
     init,
     ...props,
   });
 
-  vis.start();
+  window.visualisation.start();
 }