+++ src/renderer/GPURenderer/common/TextureAtlas/index.js
@@ -12,12 +12,13 @@ import potpack from 'potpack';
  *
  */
 export default class TextureAtlas {
-  constructor(renderer, debug) {
+  constructor(renderer, shouldDebug) {
     const { three: THREE, type: rendererType } = renderer;
     const data = new Float32Array(DATA_TEXTURE_SIZE * 4);
     const ctx = (this.ctx = document.createElement('canvas').getContext('2d'));
     const { canvas } = ctx;
 
+    this.shouldDebug = shouldDebug;
     this.rendererType = rendererType;
     this.indexData = data;
     this.canvas = canvas;
@@ -35,7 +36,7 @@ export default class TextureAtlas {
 
     canvas.width = canvas.height = DATA_TEXTURE_SIZE;
 
-    if (debug) {
+    if (shouldDebug) {
       this.debug(canvas, ctx);
     }
 
@@ -176,5 +177,23 @@ export default class TextureAtlas {
     }
 
     atlasTexture.needsUpdate = true;
+  }
+
+  /**
+   * Disposes of the textures used by the texture atlas.
+   *
+   * @return void
+   */
+  destroy() {
+    const { atlasIndex, atlasTexture, canvas } = this;
+
+    atlasTexture.dispose();
+    atlasIndex && atlasIndex.dispose();
+
+    if (this.shouldDebug) {
+      canvas.remove();
+    }
+
+    this.entries = [];
   }
 }